+++ Packages/com.unity.inputsystem/InputSystem/Controls/Processors/ScaleProcessor.cs
@@ -1,0 +1,16 @@
+namespace UnityEngine.Experimental.Input.Processors
+{
+    /// <summary>
+    /// Scale a float value by a constant factor.
+    /// </summary>
+    public class ScaleProcessor : IInputControlProcessor<float>
+    {
+        [Tooltip("Scale factor to multiply incoming float values by.")]
+        public float factor;
+
+        public float Process(float value, InputControl control)
+        {
+            return value * factor;
+        }
+    }
+}